requirements.txt
pandas 
numpy 
flask 
scikit-learn
seaborn
datetime
typing
mysql-connector
sqlalchemy
psycopg2
pymysql
-e .